app.py/dist.uv/pyproject.toml
[project]
name = "deepfungen"
version = "1.2.0"
description = ""
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "av>=10.0.0",
    "fastapi>=0.111.0",
    "jinja2>=3.1.6",
    "numpy>=1.24.0",
    "nvidia-ml-py>=13.580.82",
    "onnxruntime-directml>=1.23.0",
    "opencv-python-headless>=4.10.0",
    "pandas>=2.0.0",
    "psutil>=5.9.0",
    "pydantic>=1.10.0,<2",
    "PyTurboJPEG>=1.7.0",
    "pywebview>=4.4.2",
    "scipy>=1.15.3",
    "uvicorn[standard]>=0.30.0",
]